# swagger.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:3000/docs","description":"Community API"}],"info":{"description":"Eatbook 커뮤니티 API Docs","contact":{"email":"rkdalsdl112@gmail.com","name":"KangMin Han","url":"https://github.com/rkdalsdl98/mocat-mall-server"},"version":"1.0.0","title":"eatbook-community","license":{"name":"UNLICENSED"}},"paths":{"/novel/status":{"post":{"tags":[],"operationId":"NovelStatusController.updateReviewStatus","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body.IUpdateReviewStatusArgs"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["novel","status","updateReviewStatus"],"x-samchon-controller":"NovelStatusController"}},"/novel/list":{"get":{"tags":[],"operationId":"NovelController.getNovelList","parameters":[{"name":"page","in":"query","schema":{"type":"number","minimum":1},"required":true},{"name":"orderBy","in":"query","schema":{"oneOf":[{"const":"asc"},{"const":"desc"}]},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["novel","list","getNovelList"],"x-samchon-controller":"NovelController"}},"/novel":{"get":{"tags":[],"operationId":"NovelController.getNovel","parameters":[{"name":"id","in":"query","schema":{"type":"string","pattern":"[A-Za-z0-9]{4}-[0-9]{8,11}"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["novel","getNovel"],"x-samchon-controller":"NovelController"},"post":{"tags":[],"operationId":"NovelController.registNovel","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body.IRegistNovelArgs"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["novel","registNovel"],"x-samchon-controller":"NovelController"},"delete":{"tags":[],"operationId":"NovelController.deleteNovel","parameters":[{"name":"id","in":"query","schema":{"type":"string","pattern":"[A-Za-z0-9]{4}-[0-9]{8,11}"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["novel","deleteNovel"],"x-samchon-controller":"NovelController"}},"/requester/{email}":{"get":{"tags":[],"operationId":"RequesterController.searchRequester","parameters":[{"name":"email","in":"path","schema":{"type":"string","format":"email"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["requester","searchRequester"],"x-samchon-controller":"RequesterController"}},"/requester":{"post":{"tags":[],"operationId":"RequesterController.registRequester","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body.IRegistRequesterArgs"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{}}}},"x-samchon-accessor":["requester","registRequester"],"x-samchon-controller":"RequesterController"}}},"components":{"schemas":{"Body.IUpdateReviewStatusArgs":{"type":"object","properties":{"statusId":{"type":"string","maxLength":30},"reason":{"type":"string","maxLength":300},"status":{"type":"string"},"responsiblePersonEmail":{"type":"string","format":"email"},"responsiblePerson":{"type":"string"}},"required":["statusId","reason","status","responsiblePersonEmail","responsiblePerson"]},"QueryParam.IGetNovelList":{"type":"object","properties":{"page":{"type":"number","minimum":1},"orderBy":{"oneOf":[{"const":"asc"},{"const":"desc"}]}},"required":["page","orderBy"]},"QueryParam.IGetNovelById":{"type":"object","properties":{"id":{"type":"string","pattern":"[A-Za-z0-9]{4}-[0-9]{8,11}"}},"required":["id"]},"Body.IRegistNovelArgs":{"type":"object","properties":{"id":{"type":"string","pattern":"[A-Za-z0-9]{4}-[0-9]{8,11}"},"requester":{"$ref":"#/components/schemas/Body.IRequesterInfoArgs"},"novelInfo":{"$ref":"#/components/schemas/Body.INovelInfoArgs"}},"required":["id","requester","novelInfo"]},"Body.IRequesterInfoArgs":{"type":"object","properties":{"requesterEmail":{"type":"string","format":"email"},"requesterName":{"type":"string"}},"required":["requesterEmail","requesterName"]},"Body.INovelInfoArgs":{"type":"object","properties":{"novelTitle":{"type":"string","maxLength":200},"novelDescription":{"type":"string","maxLength":200},"ref":{"type":"string","format":"url"}},"required":["novelTitle","novelDescription","ref"]},"Body.IRegistRequesterArgs":{"type":"object","properties":{"novelId":{"type":"string","pattern":"[A-Za-z0-9]{4}-[0-9]{8,11}"},"email":{"type":"string","format":"email"},"name":{"type":"string"}},"required":["novelId","email","name"]}},"securitySchemes":{}},"tags":[],"x-samchon-emended":true}